+++ cucumber.js
@@ -1,7 +1,7 @@
 // cucumber.js
 let common = [
     '--require-module ts-node/register', // Load TypeScript module
-    '--require features/support/**/*.ts', // Load step definitions
+    '--require src/**/*.ts', // Load step definitions
     // `--format ${
     //   process.env.CI || !process.stdout.isTTY ? 'progress' : 'progress-bar'
     // }`,
