+++ src/hooks.ts
@@ -1,7 +1,8 @@
 // Dependencies
-import { After, Before, AfterAll, Status } from 'cucumber'
-import { openBrowser, closeBrowser, takeScreenshot } from './browser-actions'
-import { CentrifugeWorld } from './world'
+import { After, Before, Status } from 'cucumber'
+
+import { openBrowser, closeBrowser, takeScreenshot } from './support/browser-actions'
+import { CentrifugeWorld } from './support/world'
 
 Before(async function(this: CentrifugeWorld, scenario) {
   this.clearContext()
