+++ src/support/tinlake-actions.ts
@@ -2,7 +2,7 @@ import { ethers } from 'ethers'
 import Tinlake, { ITinlake } from '@centrifuge/tinlake-js'
 
 import { CentrifugeWorld } from './world'
-import { config } from './config'
+import { config } from '../config'
 
 export async function ensureTinlakeInit(world: CentrifugeWorld): Promise<ITinlake> {
   if (world.tinlake) {
