+++ src/utils/debug.ts
@@ -1,4 +1,4 @@
-import { CentrifugeWorld } from '../world'
+import { CentrifugeWorld } from '../support/world'
 
 export async function debug(world: CentrifugeWorld) {
   await world.currentPage.evaluate(() => {
