+++ src/utils/getTextContent.ts
